Refactor OIDC user manager initialization by extracting logic into `auth.client.ts` plugin
